+++ lib/validations.ex
@@ -9,6 +9,8 @@ defmodule ElixirCLRut.Validations do
   @doc """
   Receives a Token struct and counts the normalized array,
   if it's empty it will add :value_is_empty to the errors list.
+  iex> not_empty(ElixirCLRut.Token.from(ElixirCLRut.from("abc")))
+  %ElixirCLRut.Token{errors: [:value_is_empty], from: %ElixirCLRut.Struct{from: "abc", checkdigit: "0", lastdigit: "", lastchar: nil, normalized: [], normalized_with_checkdigit: [], includes_checkdigit?: true, dashed?: false}, valid?: false}
   """
   @doc since: "1.0.0"
   @spec not_empty(struct()) :: struct()
@@ -22,6 +24,8 @@ defmodule ElixirCLRut.Validations do
   @doc """
   Receives a Token struct and validates checkdigit with lastdigit.
   If they are different it will add :wrong_checkdigit to the errors list.
+  iex> has_valid_checkdigit(ElixirCLRut.Token.from(ElixirCLRut.from("20.961605-C")))
+  %ElixirCLRut.Token{errors: [:wrong_checkdigit], valid?: false, from: %ElixirCLRut.Struct{from: "20.961605-C", checkdigit: "0", lastdigit: "5", lastchar: nil, normalized: [2, 0, 9, 6, 1, 6, 0], normalized_with_checkdigit: [2, 0, 9, 6, 1, 6, 0, 5], includes_checkdigit?: true, dashed?: true}}
   """
   @doc since: "1.0.0"
   @spec has_valid_checkdigit(struct()) :: struct()
@@ -56,7 +60,6 @@ defmodule ElixirCLRut.Validations do
     end
   end
 
-
   @doc """
   Optional validation.
   Receives a Token struct and validates that all chars
@@ -80,6 +83,50 @@ defmodule ElixirCLRut.Validations do
     case zeroes == token.from.normalized do
       true -> Token.error(token, :all_zeroes)
       false -> token
+    end
+  end
+
+  @doc """
+  Optional validation.
+  Validates Rut is above or equal to 1 million. To avoid probably dead people.
+  - since: "1.0.2"
+
+  ## Example
+
+    iex> (ElixirCLRut.validate("1.000.000-9") |> equal_or_above_1M()).valid?
+    true
+
+  """
+  @spec equal_or_above_1M(struct()) :: struct()
+  def equal_or_above_1M(%Token{} = token) do
+    check = String.to_integer(Enum.join(token.from.normalized)) >= 1_000_000
+
+    # if we have all zeroes then is not valid
+    case check do
+      true -> token
+      false -> Token.error(token, :under_1M)
+    end
+  end
+
+  @doc """
+  Optional validation.
+  Validates Rut is below or equal to 100 million. To avoid probably non existing people.
+  - since: "1.0.2"
+
+  ## Example
+
+    iex> (ElixirCLRut.validate("100.000.000-7") |> equal_or_below_100M()).valid?
+    true
+
+  """
+  @spec equal_or_below_100M(struct()) :: struct()
+  def equal_or_below_100M(%Token{} = token) do
+    check = String.to_integer(Enum.join(token.from.normalized)) <= 100_000_000
+
+    # if we have all zeroes then is not valid
+    case check do
+      true -> token
+      false -> Token.error(token, :over_100M)
     end
   end
 end