+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Romantic Versioning](http://sentimentalversioning.org/).
 
+## [1.0.2](https://github.com/elixircl/rut/releases/tag/v1.0.2) - 2025-06-16
+### Bugs
+- Assume contains checkdigit in Ruts.
+### Features
+- Add more validation cases
+- Add above 1M and below 100M validation helpers
+
 ## [1.0.0](https://github.com/elixircl/rut/releases/tag/v1.0.0) - 2022-03-30
 ### Features
 - Can Format a _RUT/RUN_ value.